Decouple modal and toast from local RenderContext
- add new interface for *managed* components: ``ManagedComponent<T>`` This should be implemented by all components which manage their own, *internal* RenderContext like toasts and modals until now.
- integrate and rename former stand alone factory function ``globalRenderContext`` as ``managedRenderContext`` into this new interface.
- adapt ``toast`` and ``modal`` components to implement this new interface for managed components
- remove receiver from ``showToast``, ``toast`` and ``modal`` factory functions
